# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Image Upscale Service
  description: |
    A REST API for AI-based image upscaling using Real-ESRGAN.
    
    This service allows you to upload images and upscale them by 2x, 3x, or 4x using various optimized models.
  version: 1.0.0
  contact:
    name: Michael Lechner
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: .
    description: Current Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /cancel/{job_id}:
    post:
      summary: Cancel job
      description: Cancel a running or queued job.
      operationId: cancelJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Bad request (e.g. job already completed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Auth-Token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        success:
          type: boolean
          example: false